# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Assistant API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The type of task to perform
 */
export type AiQueryRequestTaskType =
  (typeof AiQueryRequestTaskType)[keyof typeof AiQueryRequestTaskType];

export const AiQueryRequestTaskType = {
  search: "search",
  code: "code",
  summarize: "summarize",
  image: "image",
} as const;

export interface AiQueryRequest {
  /** The user's query text */
  query: string;
  /** The type of task to perform */
  taskType: AiQueryRequestTaskType;
  /**
   * Optional URL for content extraction
   * @nullable
   */
  url?: string | null;
}

export interface AiQueryResponse {
  id: number;
  query: string;
  taskType: string;
  /** The consolidated AI response text */
  result: string;
  /**
   * Generated image URL (for image task type)
   * @nullable
   */
  imageUrl?: string | null;
  /** APIs that were used to generate this response */
  sources: string[];
  createdAt: string;
}

export interface HistoryItem {
  id: number;
  query: string;
  taskType: string;
  result: string;
  /** @nullable */
  imageUrl?: string | null;
  sources: string[];
  createdAt: string;
}

export type UsageStatsByTaskType = { [key: string]: number };

export interface UsageStats {
  totalQueries: number;
  byTaskType: UsageStatsByTaskType;
  recentActivity: HistoryItem[];
}

export interface ErrorResponse {
  error: string;
}

export type GetHistoryParams = {
  limit?: number;
  /**
   * @nullable
   */
  taskType?: string | null;
};