# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Assistant API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Routes the query to relevant APIs based on task type and returns consolidated response
 * @summary Send a query to the AI assistant
 */
export const AiQueryBody = zod.object({
  query: zod.string().describe("The user's query text"),
  taskType: zod
    .enum(["search", "code", "summarize", "image"])
    .describe("The type of task to perform"),
  url: zod.string().nullish().describe("Optional URL for content extraction"),
});

export const AiQueryResponse = zod.object({
  id: zod.number(),
  query: zod.string(),
  taskType: zod.string(),
  result: zod.string().describe("The consolidated AI response text"),
  imageUrl: zod
    .string()
    .nullish()
    .describe("Generated image URL (for image task type)"),
  sources: zod
    .array(zod.string())
    .describe("APIs that were used to generate this response"),
  createdAt: zod.string(),
});

/**
 * Returns recent AI query history
 * @summary Get query history
 */
export const getHistoryQueryLimitDefault = 20;

export const GetHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(getHistoryQueryLimitDefault),
  taskType: zod.coerce.string().nullish(),
});

export const GetHistoryResponseItem = zod.object({
  id: zod.number(),
  query: zod.string(),
  taskType: zod.string(),
  result: zod.string(),
  imageUrl: zod.string().nullish(),
  sources: zod.array(zod.string()),
  createdAt: zod.string(),
});
export const GetHistoryResponse = zod.array(GetHistoryResponseItem);

/**
 * @summary Delete a history item
 */
export const DeleteHistoryItemParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * Returns stats about AI usage by task type
 * @summary Get usage statistics
 */
export const GetStatsResponse = zod.object({
  totalQueries: zod.number(),
  byTaskType: zod.record(zod.string(), zod.number()),
  recentActivity: zod.array(
    zod.object({
      id: zod.number(),
      query: zod.string(),
      taskType: zod.string(),
      result: zod.string(),
      imageUrl: zod.string().nullish(),
      sources: zod.array(zod.string()),
      createdAt: zod.string(),
    }),
  ),
});