# lib/api-zod/src/generated/types/aiQueryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Assistant API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiQueryRequestTaskType } from "./aiQueryRequestTaskType";

export interface AiQueryRequest {
  /** The user's query text */
  query: string;
  /** The type of task to perform */
  taskType: AiQueryRequestTaskType;
  /**
   * Optional URL for content extraction
   * @nullable
   */
  url?: string | null;
}