# lib/api-zod/src/generated/types/aiQueryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Assistant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiQueryResponse {
  id: number;
  query: string;
  taskType: string;
  /** The consolidated AI response text */
  result: string;
  /**
   * Generated image URL (for image task type)
   * @nullable
   */
  imageUrl?: string | null;
  /** APIs that were used to generate this response */
  sources: string[];
  createdAt: string;
}