# lib/api-zod/src/generated/types/historyItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Assistant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryItem {
  id: number;
  query: string;
  taskType: string;
  result: string;
  /** @nullable */
  imageUrl?: string | null;
  sources: string[];
  createdAt: string;
}