# lib/api-zod/src/generated/types/usageStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Assistant API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryItem } from "./historyItem";
import type { UsageStatsByTaskType } from "./usageStatsByTaskType";

export interface UsageStats {
  totalQueries: number;
  byTaskType: UsageStatsByTaskType;
  recentActivity: HistoryItem[];
}